--- src/test/java/com/nucleiassignment3/HttpServer/EmployeeDaoImplTest.java
@@ -1,19 +1,0 @@
-package com.nucleiassignment3.HttpServer;
-
-import com.nucleiassignment3.HttpServer.bo.EmployeeBo;
-import com.nucleiassignment3.HttpServer.dao.EmployeeDao;
-import com.nucleiassignment3.HttpServer.dao.EmployeeDaoImpl;
-import com.nucleiassignment3.HttpServer.entity.Employee;
-import com.nucleiassignment3.HttpServer.mapper.EmployeeMapper;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Assertions.*;
-
-import java.sql.Date;
-
-public class EmployeeDaoImplTest {
-
-
-        @Test
-        void getEmployee() {
-        }
-    }
